refactor: Refactor Pipeline Services
